Refactor/replace font support with typography

+++ src/com/hridoy/thmlt/utility/TypographyTemplate.java
@@ -3,10 +3,10 @@ package com.hridoy.thmlt.utility;
 public class TypographyTemplate {
     private final int fontSize;
     private final int lineHeight;
-    private final int letterSpacing;
+    private final float letterSpacing;
     private final String linkedFont;
 
-    public TypographyTemplate(int fontSize, int lineHeight, int letterSpacing, String linkedFont) {
+    public TypographyTemplate(int fontSize, int lineHeight, float letterSpacing, String linkedFont) {
         this.fontSize = fontSize;
         this.lineHeight = lineHeight;
         this.letterSpacing = letterSpacing;
@@ -15,7 +15,7 @@ public class TypographyTemplate {
 
     public int getFontSize() { return fontSize; }
     public int getLineHeight() { return lineHeight; }
-    public int getLetterSpacing() { return letterSpacing; }
+    public float getLetterSpacing() { return letterSpacing; }
     public String getLinkedFont() { return linkedFont; }
 }
